Remove mask_provider from horizontal coords (#1982)
This PR removes `mask_provider` as item within `LatLonCoordinates`.
Instead, `mask_provider` is an argument when generating
`GriddedOperations` from `LatLonCoordinates`. This has a couple
benefits. 1) now there is a single source of truth for the mask provider
in dataset properties/info and 2) we don't have to worry about
serializing `mask_provider` within horizontal coordinates which was
otherwise a pain point (see #1979).

Changes:
- remove `mask_provider` from `LatLonCoordinates`
- update call signature of `HorizontalCoordinates.gridded_operations` so
that it takes an optional mask provider (default is NullMaskProvider) in
order to generate/return the gridded operations
- `DatasetInfo.gridded_operations` now appropriately combines the
horizontal coordinate and mask provider
- refactor aggregators to initialize using dataset info instead of
individual coordinates/timestep

- [x] Tests added