Fix weight freezing when using StepABC (#1994)
This PR fixes an issue where using frozen weights will result in an
exception if running in parallel, because DDP is applied before the
weights are frozen and expects them to have gradients.

Changes:
- Added `init_weights` callable argument to `StepABC.get_step` which
initializes (is used to freeze) model weights after model definition but
before DDP is wrapped over the module.

- [x] Tests added