Implement placeholder weather evaluation (#1995)
We would like to be able to evaluate weather skill using runs configured
separately from long inference skill, and using different metrics. For
example, we would like to have a larger ensemble potentially with
multiple ensemble members per IC (and report weather skill metrics) for
weather evaluation, while having fewer ensemble members during long
inference and ignoring their weather metrics (which are not robust).

This PR only adds the weather evaluation as a copy of the current inline
inference. Later PRs will modify it to be specialized on weather, and to
remove weather metrics from the current inference.

Changes:
- Added `weather_evaluation` section to TrainConfig which is configured
identically to InlineInferenceConfig, allowing a second form of inline
inference to be run.
- Added `Slice.contains` helper function to determine if the slice
contains an index without constructing an explicit list of indices.

- [x] Tests added