Parallelize `_get_raw_times` when loading from many files (#2033)
@mcgibbon has observed that `_get_raw_times` is a bottleneck when
initializing an `XarrayDataset`. It currently iterates in serial over
the `paths` it is provided, which could be quite slow in the case that
we are loading from monthly netCDFs, i.e. O(1000) files.

This PR parallelizes this with a `multiprocessing.Pool` when the number
of paths to load is exceeds 12. This is a relatively arbitrary cutoff,
mainly to ensure that tests, which run this code in quick succession on
a small number of files, do not need to worry about the overhead of
setting up a `Pool`.

I suspect opening many files is the bottleneck rather than time decoding
itself (i.e. this should already not be a bottleneck when we load from
zarr).

Changes:
- symbol (e.g. `fme.core.my_function`) or script and concise description
of changes or added feature
- Can group multiple related symbols on a single bullet

- [x] Tests added