Pull torch.no_grad context above inference (#2041)
This PR removes the `torch.no_grad` context call from within the
inference helper, placing it above instead. This allows the helper to be
called in a case where we do want to obtain gradients, such as to
compute jacobians in a notebook.

Changes:
- Removed the `torch.no_grad` context call from within the inference
helper, placing it above in our entrypoints instead
- Inference DataWriters updated to accept tensors which require
gradient.

- [ ] Tests added