Avoid stopiteration exception on iter call (#2054)
There is a bug where if our model gets restarted at the end of an epoch,
during validation or inference, the data loader will have no batches
when restarted. This would normally be OK, but our
SlidingWindowDataLoader tries to get the next batch during an `__iter__`
call, where StopIteration is a normal exception (it only gets treated as
the end of an iteration during `__next__`. This PR resolves that issue
by returning an empty tensor iterator instead.

Changes:
- Fixed an exception which occurs if the model is restarted at the end
of an epoch, during validation or inference, if buffer timesteps are
used.

- [ ] Tests added