Add a regional ACE data writer for downscaling outputs. (#2038)
Add a new data writer option for ACE inference that allows users to
output data from regions at defined times.

Changes:
- `SubselectWriterConfig` used to define the output writer for a single
region, used to produce a `SubselectWriter`
- Added a new section to the `DataWriterConfig`, `subselection` which
accepts a list of `SubselectWriterConfig` to define desired of new
writer outputs with subselected time or regions
- `TimeSlice` in `fme.core.dataset.time` now has a `as_raw_slice` method
to return the slice with the date strings (directly usable by xarray)
- `Slice` in `fme.core.typing_` has a `shift_left` method which returns
a new `Slice` with shifted bounds based on the `start_index`. This
allows for specifying an absolute integer slice that is still respected
across individual inference batches.

- [x] Tests added