Fix one step map when record_batch is given multiple lead times (#2071)
Currently the one step MapAggregator will raise an exception if given
data with multiple timeseries lengths, even though it only uses the
first step predicted output. This PR updates it so it no longer raises
an exception on this case, by selecting the time index at record time
instead of at final reporting time. This also reduces the memory
footprint of the aggregator.

A test is not included in this PR, but this fails during tests in
`feature/stepper_multiple_lead_times` which is in the process of being
merged.

Changes:
- No API or behavioral changes

- [ ] Tests added