Add data processing configuration for one degree X-SHiELD data (#2074)
This PR adds a configuration for processing the 1° Gaussian grid
X-SHiELD data for emulation. The configuration selects the same
variables to those processed in our existing SHiELD-AMIP configurations.

The dataset spans `2013-01-01T06:00:00` to `2024-01-04T00:00:00`
(inclusive on both sides). We might consider the first year as spin-up,
so I configured the stats computation to use the time range
`2014-01-01T06:00:00` to `2024-01-01T00:00:00` (inclusive on both
sides), but we could consider changing this.

Argo workflow: `compute-fme-dataset-ensemble-whz9h`.