Rename near-surface variables in X-SHiELD data processing as in other configs (#2081)
@Arcomano1234 noted that in #2074 we forgot to apply the same `renaming`
config to the near-surface variables in the X-SHiELD data processing
config that we apply in other configs, e.g. in the original SHiELD-AMIP
config:
https://github.com/ai2cm/full-model/blob/d890b5add272fae90c76fb59c418e176929dbff6/scripts/data_process/configs/shield-amip-ensemble-c96-1deg-8layer.yaml#L24-L28

This updates it to be consistent. I resubmitted the processing job
targeting a new path for the data and stats. We can delete the old data
once this finishes. This is the argo job:
`compute-fme-dataset-ensemble-mcdtt`.