Add labels to BatchData (#2085)
This PR adds labels to BatchData, which will be used in the future for
domain embeddings.

Changes:
- Added `labels: list[str]` configuration option to `XarrayDataConfig`,
allowing the user to specify labels to assign to its batches
- Added `labels: set[str]` attributes to BatchData, containing the
labels of its dataset if present.
- Added `labels: set[str]` and `horizontal_dims: list[str]` attributes
to PairedData, so its data aligns with what is available from BatchData
- Added `labels: set[str]` option to InferenceConfig, which will
override dataset labels if present. This ensures the initial condition
and forcing labels are compatible.

- [x] Tests added