Allow support for multiple training lead times (#2050)
This PR adds the ability to train with a stochastic number of forward
steps for each batch.

Currently data for each batch is loaded for the maximum number of
forward steps, as the number of steps is difficult to define at the time
when the data is loaded. This could potentially be improved in a later
PR by defining at each epoch the number of forward steps in each index
of the XarrayDataset, but it requires determining which indices will end
up in batches with one another, which is challenging. It does appear
training can still be performant with this downside, as the additional
overhead to retrieve time-contiguous samples is not as much as
retrieving non-contiguous samples.

Changes:
- Deprecated `TrainConfig.n_forward_steps` in favour of
`StepperConfig.train_n_forward_steps`
- Added TimeLengthProbabilities type for train_n_forward_steps, which
allow specifying multiple potential lead times with associated
probability weights

- [x] Tests added