Update era5 baseline with qol features (#2131)
This PR updates the ERA5 baseline training script with quality-of-life
features @Arcomano1234 and I have been using in our experiments. Not all
of these actually get used by the baseline, but a primary purpose of the
baseline is as a copy-paste starting point for experiments. Having these
features there will make it easier to take advantage of them in
experimentation.

Eventually we may want to centralize these features in a way that's
easier to re-use instead of needing to copy-paste between experiments.
That is not done here.

Changes:
- Moved ERA5 stats dataset into its ace-train-config.yaml instead of the
submission `.sh` script. This allows a single experiment directory to be
used with multiple stats datasets across multiple configs.
- Allow a `$WANDB_USERNAME` environment variable to be set, overriding
the beaker username default. This is needed for users whose wandb and
beaker usernames are different.
- Use a `run_training` helper function to allow multiple job submissions
to occur in the same `.sh` script without copy-pasting logic.

- [ ] Tests added