Port wandb_in_experiment_dir option to public codebase (#42)
This PR allows toggling whether the `wandb` directory will be stored in the `experiment_dir` or in `/tmp`. By default it retains the existing behavior, where it is stored in local `/tmp`.  Storing it in the `experiment_dir` is useful when running bare-metal, since it ensures that the directory is unique and can be accessed from nodes that have internet connectivity.

Changes:
- Adds a `wandb_in_experiment_dir` option (default `False`) to the `LoggingConfig` to toggle whether the `wandb` directory will be stored in the experiment directory or not.