Add a `finalize` method to data writers to close files (#2190)
This PR adds a `finalize` method to each of the data writers to ensure
written files are appropriately flushed and closed at the end of
inference and evaluation. In some circumstances we need to make sure we
call this within tests of the writers to ensure we can open their
resulting files with xarray later.

Changes:
- Adds a `finalize` method to the data writers, which flushes the
dataset buffer and then closes the dataset.
- Removes upper bound on xarray version temporarily introduced in #2191.

Resolves #2188.