Fix missing nino34 index power spectrum (#2175)
We added the capability of logging the power spectrum of nino34 index,
but it was missing for some specific runs. This PR fixes this issue, it
has to do with inline inference not having overlapping time period since
we dropnan along time to calculate power spectrum.

Changes:
- When calculating nino34 index power spectrum, we now truncate to the
shortest timeseries length of all samples

Resolves #1846

---------

Co-authored-by: Oliver Watt-Meyer <oliverwatt@gmail.com>