Fix bug in passing `channel_mean_names` to `TimeMeanEvaluatorAggregator` (#2203)
Fixes a minor bug where `InferenceEvaluatorAggregator`'s
`channel_mean_names` init argument wasn't passed on to
`TimeMeanEvaluatorAggregator`.

Note that there is effectively no change in the channel-mean computation as a result
of this bugfix since the default behavior of
`TimeMeanEvaluatorAggregator` with `channel_mean_names=None` is to
compute the channel-mean RMSE of all the generated outputs.

- [x] Tests added