Add normalize_big_skip option to NoiseConditionedSFNO (#2147)
This PR adds an option normalize_big_skip which if given, applies
LayerNorm to the input before passing it to the decoder when big_skip is
enabled. This ensures every path from the input to the output of the
network passes through a LayerNorm, which may help with generalization.

Changes:
- Added `normalize_big_skip` option to NoiseConditionedSFNO

- [x] Tests added