Rename `SubselectWriter` to `FileWriter` (#2230)
Since the `SubselectWriter` can be used for more than just subselection,
and is more flexible than the existing data writing configuration style,
we plan to deprecate existing style and require configuration through a
list of single file writer configs. This rename clarifies that more
general purpose nature of the `SubselectWriter` which is now renamed
`FileWriter` (and similar for the corresponding configs).

Changes:
- rename `SubselectWriter` -> `FileWriter` and `SubselectWriterConfig`
-> `FileWriterConfig`

---------

Co-authored-by: Anna Kwa <annak@allenai.org>