Update README with links to refereed manuscripts (#2236)
This PR updates the README with consistent titles and links to the
refereed versions of the ACE2-SOM and Green's Function manuscripts. It
is a copy of ai2cm/ace#43.

I think it might make sense to merge this first and then in a follow-up
PR make whatever changes to the README we would like to reference
SamudrACE and the associated paper.

cc: @brianhenn