Improved FileWriter file names (#2271)
Improves the file names produced by the `FileWriter` via the
`FileWriterConfig`. Pre-existing data writer API is not changed.

Changes:
- Only if `FileWriterConfig.save_reference` is `True` are `_predictions`
and `_target` suffixes are appended to the relevant files of the
`PairedFileWriter` (as previously); if `FileWriterConfig.save_reference`
is `False` no suffixes are appended as only predictions are written out.
- If `FileWriterConfig.time_coarsen` is `MonthlyCoarsenConfig`,
`monthly_mean_` is no longer prefixed the the filename automatically.
- A validation check is added to ensure that duplicate filenames are not
chosen via the `FileWriterConfig`.

- [X] Tests added