Downscaling cascade evaluation fix (#2245)
This PR adds a small adjustment to fix an issue where each step of the
cascaded predictor was using the top-level batch to grab data keys to
create an example current sample batch. In the failure case, the
top-level batch had variables not predicted by a previous model. This
adjusts to only ever grab sample data from the current batch instead of
using the top-level keys.