Implement ability to compute insolation as in CM4 (#2250)
This PR is the first of a sequence of two to enable computing insolation
dynamically at training and inference time in ACE. As a start it
implements the method used in CM4, but is written to enable
straightforwardly hooking in other implementations.

Changes:
- Adds an `insolation` subdirectory within `fme/ace/stepper` for code
related to configuring and computing insolation.
- Adds a `fme.ace.stepper.insolation.cm4` module with an implementation
of the method for computing interval-averaged insolation in CM4. For now
this is limited to time intervals up to 12 hours in length per their
algorithm. This required some augmenting of the test code to facilitate
writing test data with a timestep less than 12 hours.
- Adds a `InsolationConfig` for configuring parameters for computing
insolation (currently only as in CM4), as well as `NameConfig` and
`ValueConfig` sub-configs for configuring the solar constant to be
loaded from disk or set as a constant scalar value.

- [x] Tests added

See #2279 for the PR that integrates this into the `Stepper`. [This
notebook](https://github.com/ai2cm/explore2/blob/main/spencerc/2025-09-10-derived-insolation/2025-10-31-CM4-insolation-implementation-validation.ipynb)
shows that the implementation here matches the regridded output from CM4
for year 0300 to within an RMSE less than 0.65 W/m<sup>2</sup> at each
time interval.