Enable computing derived forcings within the `Stepper` (#2279)
This PR introduces the ability to compute derived forcings within the
`Stepper`, with the initial application of computing derived insolation.
It builds upon #2250, so targets that branch for clarity of diff.

I ran a couple one-epoch production training jobs on the CM4 dataset
with and without derived insolation to assess the impact of computing
insolation on throughput. It does not seem to make a meaningful
difference based on the `epoch_training_seconds` and
`epoch_inference_seconds` metrics ([WandB
view](https://wandb.ai/ai2cm/ace2-cm4/workspace?nw=xuuezo6q0vb)).

Changes:
- Adds a `derived_forcings` option to the `StepperConfig`, which accepts
a `DerivedForcingsConfig` object.
- Adds `DerivedForcingsConfig` and `ForcingDeriver` objects to handle
configuring and computing derived forcings.
- Integrates derived forcings into the `Stepper` to enable computing
derived forcings during training and inference, and serializing
configuration information with trained models.
- The `derived_forcings` configuration of a serialized `Stepper` can be
overridden via the `StepperOverrideConfig`. In the case of insolation
one can only override the configuration with one with the same
`insolation_name`.

- [x] Tests added