Rename 2d to noise in conditioning (#2292)
The model conditioning currently uses a generic "2d" conditioning tensor
for noise conditioning. If an additional type of 2d conditioning were
added to this, it would only be possible to modify the size of the last
type in the tensor, which is confusing API-wise. This PR renames "2d" to
"noise" conditioning so that if any new conditioning types are added,
they are to be put into their own names.

Changes:
- No public API changes

- [ ] Tests added