Generation entrypoint: zarr chunking utilities (#2298)
## Automated Zarr Chunking that works with parallel generation

I am breaking up the generation entrypoint PR (#2265) into a few PRs for
ease of review. The first one just handles the utilities used for
defining zarr chunks for the base case of when the user does not provide
any chunk specification.

The utility will take a given shape (the maximum output shape from a
single unit of GPU generation), and give a chunk shape for the output
zarr. The utility ensures reasonable size (defaults to <=10MB) by
dividing the spatial dimensions [-2, -1], if necessary. The code assumes
that time/ens dimension chunks will be 1, such that they will evenly fit
into a shard determined by the work item size.

## design principle

The idea of determining this automatically is with the hope that the end
user shouldn't have to think about chunking, they just get a reasonable
output, and they also don't have to think about whether that works for
parallel generations. We still have the ability to specify chunks in
configuration, but this will be noted as an advanced "sharp" feature
that's easy to get wrong for parallel distribution of zarr writes.

Note that GPU work is never divided among the spatial dimensions
(lat/lon), only the time and ensemble dimension. That is, each unit of
GPU work will provide the fully requested spatial field for a number of
times and ensemble members. This will comprise of a single object write
output as a shard, unless the shard is the same size as the underlying
chunk, in which case, we will only provide a chunk (not handled in this
PR).

### Example 

Say we want to downscale to produce a single time 20-member ensemble
over a region. The user has a GPU batch size option, controlling how
much memory and time each unit of computation takes. In this case, a GPU
batch size of 15, produces a unit of downscaled output that is at a
maximum (1, 15, 512, 512). In this case there would be two work items to
be handled. (1, 15, 512, 512) and (1, 5, 512, 512).

The spatial output (512, 512) is already < 10MB, so no reduction is
necessary, the zarr chunk would be (1, 1, 512, 512). The maximum work
item output shape would be (1, 15, 512, 512), which would be the zarr
shard (a single output write unit). In the next PR, load balancing might
shift how we use gpu batch size to set the work item size, but I'll
address in the next PR.

Changes:
- `determine_zarr_chunks` handles defining chunks from a given input
shape

- [x] Tests added

---------

Co-authored-by: Anna Kwa <annak@allenai.org>