Generation entrypoint: work item definition and dataset (#2299)
This PR is a continuation of the series of PRs breaking up the
generation entrypoint (#2265). This one incorporates the `SliceWorkItem`
concept that will provide a GPU with a batch of data to generate a
specific number of ensemble members.

Changes:
- `SliceWorkItem` and `LoadedSliceWorkItem` combine information of a
slice for time, and ensemble, with an associated `BatchData`. This will
be used to generate and then insert into a zarr using `ZarrWriter`
- `SliceItemDataset` creates a torch dataset that will provide work
items in single or distributed sessions to be processed. This dataset is
always processed with batch_size=1 since it's more of a "meta" dataset
where the act of accessing items (e.g., `dataset[0]`) will load the
underlying `BatchData`.
- `get_work_items` is used as a simple way to split up a total ensemble
and time range into unique set of slices and create the work items. With
some load balancing in the distributed case to prevent large mismatches
in computation between individual slices.

- [x] Tests added