Generation entrypoint: add output targets for downscaling configuration (#2300)
This PR is a continuation of the series of PRs breaking up the
generation entrypoint (https://github.com/ai2cm/full-model/pull/2265).
This PR includes the `OutputTarget` definition and configurations for
different types of targets (`EventConfig` and `TimeRangeConfig`),
similar to what's used in `evaluator.py` and `predict.py` used to define
specific places and times to downscale.

Using these configurations to resolve into a single data type helps
consolidate how we then use these targets downstream in the generation
loop.

As discussed w/ @AnnaKwa, one remaining task is how do we also make sure
these are compatible with our aggregation, but this feature is not
necessary for an initial release.

The big question is whether aggregators can handle being called to
record partial ensembles for specific times and still maintain usage of
reduced metrics. If we can, we can tie aggregation instantiation
specific to each output target configuration type and then have it work
equivalently when we store it in an `OutputTarget`, I think we'd be in a
good spot.

Changes:
- `DownscalingOutput` class defines the time/time range and region that
will be used to generate "work" items to be processed by a downscaling
model
- `DownscalingOutputConfig` base class, and specific `EventConfig` and
`TimeRangeConfig` for building output targets. We provide a minimal set
of required inputs that will overlay with a default data loader
configuration. There are options to override the underlying dataset or
patching method rather than overlaying with the default configuration

- [x] Tests added