Generation entrypoint: add generation files and cli entrypoint (#2301)
This PR is a continuation of the series of PRs breaking up the
generation entrypoint (https://github.com/ai2cm/full-model/pull/2265).
This is the final in the series and ties together the runtime with a
top-level configuration `GenerationConfig`, a `Downscaler`, and the
`main` entrypoint.

Changes:
- finished downscaling submodule and entrypoint under
fme.downscaling.inference
- CLI `python -m fme.downscaling.inference config.yaml`, also supports
DDP via `torchrun`
- `Downscaler` orchestrates the running of each output target specified
via configuration
- `GenerationConfig` ties together the entire configuration for
downscaling generation

- [x] Tests added

---------

Co-authored-by: Anna Kwa <annak@allenai.org>