Datasets conform to DatasetABC (#2325)
This PR adds a DatasetABC which our datasets conform to, so that it is
easier to extend their interface.

Note some methods are not implemented and raise exceptions, due to a
long-standing issue where we use the same classes for training and
inference datasets.

This is necessary for adding a `set_epoch` method in a later PR.

Changes:
- Added DatasetABC
- Implemented missing DatasetABC features on datasets
- Replaced `total_timesteps` count on XarrayDataset with
`validate_n_forward_steps` method on DatasetABC

- [ ] Tests added