Compute train metrics at end of epoch (#2282)
This PR updates the trainer so that it calls the TrainAggregator on
model output on training data at the end of the epoch.

The data shuffling from the next epoch is used so that the aggregator is
not evaluating specifically the oldest samples from the epoch, but
rather a fresh shuffle.

Changes:
- Added `train_evaluation_samples` argument to TrainConfig, which
controls how many samples (not batches) are considered by the
TrainAggregator
- Added `stop_batch` argument to `subset` methods on data loaders
- Renamed `start` to `start_batch` on data loader methods to make it
clearer what the integer refers to
- Added `train_evaluation_batches` attribute to the config ABC required
by the generic Trainer

- [x] Tests added

---------

Co-authored-by: Troy Arcomano <troya@allenai.org>