Condition stochastic SFNO on labels (#2338)
This PR adds the ability to condition models on labels. This is
implemented for the Conditional SFNO.

Changes:
- Updated BatchData.broadcast_ensemble to require the ensemble being
broadcast has only 1 member
- Added `n_labels` argument to the ModuleConfig ABC build method and all
implementations
- Added `labels` argument to StepABC.step
- Added `Module` wrapper class which handles label encoding.
- Updated BatchLabels from `list[set[str]]` to a custom type
- None is now provided when labels are not being used, allowing
disambiguation from the case that is using labels but has none.
- LabelEncoder is replaced with LabelEncoding, and describes the way a
given BatchLabels is encoded, allowing labels to be transitioned between
encodings.
- StepABC.step now takes a StepArgs container instead of individual
low-level arguments.
- Fixed a bug where MultiCallStep would fail if the forcing being
multi-called were in the next-step forcing names.

- [x] Tests added