Add atmosphere processing configuration for CM4-like AM4 random CO2 ensemble (#2353)
This PR adds an atmosphere data processing configuration and `make`
target for the CM4-like AM4 random CO2 ensemble. We process the full
extent of each ensemble member, and the exclude the spin-up period of
each (the first three months) from the computation of the statistics.

Successful `argo` job: `compute-fme-dataset-ensemble-tz7d5`