Samudra simple checkpointing (#2370)
We implement two checkpointing strategies at the Samudra module level
(simple and all, thanks to discussion with OpenAthena). While we allow a
version of "all" checkpointing in optimization, it is automatically
applied to both atmosphere and ocean during coupled training. Here, we
want to apply to ocean only. In the future, we can look into how to
harmonize this better.

**High level summary**
- simple method saves ~3% memory
- all method saves ~5% memory

while both have a similar GPU penalty (~5%). See
[slide](https://docs.google.com/presentation/d/1T_5-zYH-0l6AVo-r25masuA-omkjlf0dqeqgd9V0D-M/edit?slide=id.p#slide=id.p)
for more information.