Dynamically extend lead time dimension in `MonthlyDataWriter` (#2372)
This PR fixes a known issue in the monthly data writer, which is that it
currently attempts to estimate the size of the lead time dimension with
a heuristic at initialization time, which is inexact and breaks down for
long simulations and those with non-standard calendars.

Here we convert the lead time dimension to an "unlimited" dimension and
extend it dynamically as needed as an inference run proceeds. This
guarantees that the lead time dimension size will always be consistent
with the number of monthly time bins.

Changes:
- Converts the lead time dimension in `MonthlyDataWriter.dataset` to an
unlimited dimension, which is dynamically extended as needed within
`MonthlyDataWriter.append_batch`.
- Removes the `n_months` argument in the `MonthlyDataWriter`
constructor.
- Removes the heuristic `months_for_timesteps` function, which is no
longer needed.
- Writes a `"calendar"` attribute for the `"init_time"` and
`"valid_time"` coordinates, which ensures that they will be decoded with
the proper calendar downstream.

- [x] Tests added

Resolves #545
Resolves #1246
Resolves #2358
Resolves #2371