Sea ice corrector (#2367)
This is a corrector for the `FloeNet` sea ice emulator architecture. The
corrector ensures that the sea ice concentration, sea ice mass, and snow
mass are equal to the sum of the predicted source, sink, and export
budget terms for each variable (mass conserving). Closing this budget
also ensures that sea ice concentration is bounded between [0,1] and
that sea ice mass and snow mass are greater than or equal to 0. The
budget also includes a masking strategy, which sets the derived sea ice
concentration and snow mass to 0, wherever the derived sea ice mass is
0.

Because the sea ice training data do not have a vertical coordinate, the
corrector is implemented into the `NullVerticalCoordinate` class.
Therefore, this PR makes 2 file changes:

- fme.core.coordinates
- fme.core.corrector.ice